fountain-demote-to-recipe: s1
