+++ examples/recipes/gpu-particles/tsconfig.json
