--- examples/showcase/fountain/package.json
@@ -1,17 +1,0 @@
-{
-    "name": "fountain",
-    "version": "0.0.0",
-    "private": true,
-    "type": "module",
-    "scripts": {
-        "test": "bun test --cwd ../../.. ./examples/showcase/fountain/src/fountain.test.ts",
-        "gate": "playwright test"
-    },
-    "dependencies": {
-        "@dylanebert/shallot": "workspace:*",
-        "typegpu": "~0.12.0"
-    },
-    "devDependencies": {
-        "@playwright/test": "^1.58.2"
-    }
-}
